Split out validation, update references, add static validate
